feat: insight chart
