+++ apps/cloud/src/app/features/home/insight/copilot.ts
@@ -1,3 +1,4 @@
+import { getChartType } from '@metad/core'
 import {
   C_MEASURES,
   ChartAnnotation,
@@ -9,65 +10,8 @@ import {
   assignDeepOmitBlank
 } from '@metad/ocap-core'
 import { tryFixDimension } from '@metad/story/story'
-import { getChartType } from '@metad/story/widgets/analytical-card'
 import { cloneDeep, upperFirst } from 'lodash-es'
 import { z } from 'zod'
-import { ChartTypes } from './types'
-
-export const ChartSchema = z.object({
-  cube: z.string().describe('The cube name used by the chart'),
-  chartType: z.object({
-    type: z.enum(ChartTypes).describe('The chart type'),
-    chartOptions: z
-      .object({
-        seriesStyle: z.any().describe('The series options of ECharts library'),
-        legend: z.any().describe('The legend options of ECharts library'),
-        axis: z.any().describe('The axis options of ECharts library'),
-        dataZoom: z.any().describe('The dataZoom options of ECharts library'),
-        tooltip: z.any().describe('The tooltip options of ECharts library')
-      })
-      .describe('The chart options of ECharts library')
-  }),
-  dimensions: z
-    .array(
-      z.object({
-        dimension: z.string().describe('The name of dimension'),
-        hierarchy: z.string().optional().describe('The name of the hierarchy in the dimension'),
-        level: z.string().optional().describe('The name of the level in the hierarchy')
-      })
-    )
-    .describe('The dimensions used by the chart'),
-  measures: z
-    .array(
-      z.object({
-        measure: z.string().describe('The name of the measure'),
-        order: z.enum(['ASC', 'DESC']).optional().describe('The order of the measure'),
-        chartOptions: z.any().optional().describe('The chart options of ECharts library')
-      })
-    )
-    .describe('The measures used by the chart; At least one measure'),
-  slicers: z
-    .array(
-      z.object({
-        dimension: z
-          .object({
-            dimension: z.string().describe('The name of the dimension'),
-            hierarchy: z.string().optional().describe('The name of the hierarchy in the dimension'),
-            level: z.string().optional().describe('The name of the level in the hierarchy')
-          })
-          .describe('The dimension of the slicer'),
-        members: z
-          .array(
-            z.object({
-              value: z.string().describe('the key of the member'),
-              caption: z.string().describe('the caption of the member')
-            })
-          )
-          .describe('The members in the slicer')
-      })
-    )
-    .describe('The slicers used by the chart')
-})
 
 export const SuggestsSchema = z.object({
   suggests: z.array(z.string().describe('The suggested prompt')).describe('The suggested prompts')
@@ -109,6 +53,9 @@ export function transformCopilotChart(answer: any, entityType: EntityType) {
   }
 
   const dimensions = (answer.dimension ? [answer.dimension] : answer.dimensions) ?? []
+  if (dimensions.length === 0) {
+    throw new Error('At least one dimension is required.')
+  }
   chartAnnotation.dimensions = dimensions.map((dimension) => {
     return {
       ...dimension,
@@ -124,6 +71,9 @@ export function transformCopilotChart(answer: any, entityType: EntityType) {
   })
 
   const measures = answer.measure ? [answer.measure] : answer.measures ?? []
+  if (measures.length === 0) {
+    throw new Error('At least one measure is required.')
+  }
   chartAnnotation.measures = measures.map(
     (measure) =>
       ({
