--- apps/cloud/src/app/features/home/insight/types.ts
@@ -1,1 +1,0 @@
-export const ChartTypes = ['Bar', 'Column', 'Pie'] as const
