+++ libs/core-angular/src/lib/copilot/cube.schema.ts
@@ -1,0 +1,3 @@
+export function makeCubeRulesPrompt() {
+  return `The dimensions consist of three attributes: dimension, hierarchy, and level, each of which is taken from the name of dimension, hierarchy, and level in the cube, respectively.`
+}
