+++ libs/core-angular/src/lib/copilot/index.ts
@@ -2,3 +2,5 @@ export * from './indicator.schema'
 export * from './actions'
 export * from './utils'
 export * from './types'
+export * from './cube.schema'
+export * from './chart.schema'
