+++ libs/story-angular/src/lib/copilot/story/chat.ts
@@ -1,7 +1,8 @@
 import { CopilotChatMessageRoleEnum, getFunctionCall } from '@metad/copilot'
 import { calcEntityTypePrompt } from '@metad/core'
 import { omitBlank, pick } from '@metad/ocap-core'
-import { StoryPoint, StoryPointType, StoryWidget, WidgetComponentType, fixDimension, uuid } from '@metad/story/core'
+import { StoryPoint, StoryPointType, StoryWidget, WidgetComponentType, uuid } from '@metad/story/core'
+import { fixDimension } from '@metad/story/story'
 import { nanoid } from 'ai'
 import { combineLatest, concat, of } from 'rxjs'
 import { map, switchMap, tap } from 'rxjs/operators'
