+++ libs/story-angular/src/lib/explorer/explorer.component.ts
@@ -26,13 +26,13 @@ import { MatTabsModule } from '@angular/material/tabs'
 import { MatRadioModule } from '@angular/material/radio'
 import { MatDialog, MatDialogModule, MatDialogRef } from '@angular/material/dialog'
 import { PropertyCapacity, PropertyModule } from '@metad/components/property'
-import { NxChartType, ThemeService, nonBlank } from '@metad/core'
+import { CHARTS, NxChartType, ThemeService, getChartType, nonBlank } from '@metad/core'
 import { AnalyticalCardModule } from '@metad/ocap-angular/analytical-card'
 import { AnalyticalGridModule } from '@metad/ocap-angular/analytical-grid'
 import { NgmMemberTreeComponent } from '@metad/ocap-angular/controls'
 import { DisplayDensity, NgmDSCoreService, OcapCoreModule } from '@metad/ocap-angular/core'
 import { EntityCapacity, NgmEntityPropertyComponent, NgmEntitySchemaComponent } from '@metad/ocap-angular/entity'
-import { CHARTS, NgmChartPropertyComponent, NgmChartSettingsComponent, NgmFormlyChartTypeComponent, getChartType } from '@metad/story/widgets/analytical-card'
+import { NgmChartPropertyComponent, NgmChartSettingsComponent, NgmFormlyChartTypeComponent } from '@metad/story/widgets/analytical-card'
 import { NgmGridSettingsComponent } from '@metad/story/widgets/analytical-grid'
 import {
   C_MEASURES,
